Removed bottle server and renamed some files
